feat(react): separated controls into it's own component.
