+++ packages/react-player/src/react-player.tsx
@@ -4,11 +4,11 @@
 
 import type { RendererSettings, PlayMode } from 'common';
 import { PlayerState } from 'common';
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { createContext, useEffect, useMemo, useRef } from 'react';
 import type { MutableRefObject } from 'react';
 
 import { useDotLottiePlayer } from './hooks/use-dotlottie-player';
-import type { DotLottieRefProps } from './hooks/use-dotlottie-player';
+import type { DotLottieRefProps, UseDotLottiePlayerReturn } from './hooks/use-dotlottie-player';
 
 export enum PlayerEvents {
   Complete = 'complete',
@@ -35,8 +35,7 @@ export interface DotLottiePlayerProps extends React.HTMLAttributes<HTMLDivElemen
   loop?: number | boolean;
   lottieRef?: MutableRefObject<DotLottieRefProps | undefined>;
   mode?: PlayMode;
-  // TODO: properly handle unknown.
-  onEvent?: (name: PlayerEvents, params?: unknown) => void;
+  onEvent?: <T extends PlayerEvents>(name: T, params?: unknown) => void;
   playOnHover?: boolean;
   renderer?: 'svg' | 'canvas' | 'html';
   rendererSettings?: RendererSettings;
@@ -44,6 +43,17 @@ export interface DotLottiePlayerProps extends React.HTMLAttributes<HTMLDivElemen
   src: Record<string, unknown> | string;
   testId?: string;
 }
+
+export interface DotLottieContextProps extends UseDotLottiePlayerReturn {
+  loop: boolean;
+}
+
+export const DotLottieContext = createContext<DotLottieContextProps>({
+  currentState: PlayerState.Initial,
+  seeker: 0,
+  frame: 0,
+  loop: false,
+});
 
 export const DotLottiePlayer: React.FC<DotLottiePlayerProps> = ({
   onEvent,
@@ -63,11 +73,10 @@ export const DotLottiePlayer: React.FC<DotLottiePlayerProps> = ({
   src,
   className,
   testId,
+  children,
   ...props
 }) => {
   const container = useRef(null);
-  const [_prevState, setPrevState] = useState(PlayerState.Loading);
-  const [isLoop, setIsLoop] = useState(loop);
 
   const { currentState, dotLottiePlayer, frame, seeker } = useDotLottiePlayer(src, container, {
     lottieRef,
@@ -80,7 +89,7 @@ export const DotLottiePlayer: React.FC<DotLottiePlayerProps> = ({
       ...rendererSettings,
     },
     hover: playOnHover,
-    loop: isLoop,
+    loop,
     direction,
     speed,
     intermission,
@@ -89,78 +98,12 @@ export const DotLottiePlayer: React.FC<DotLottiePlayerProps> = ({
     testId,
   });
 
-  const isPlaying = useMemo(() => {
-    return currentState === PlayerState.Playing;
-  }, [currentState]);
-
-  const isPaused = useMemo(() => {
-    return currentState === PlayerState.Paused;
-  }, [currentState]);
-
-  const isStopped = useMemo(() => {
-    return currentState === PlayerState.Stopped;
-  }, [currentState]);
-
   const isError = useMemo(() => {
     return currentState === PlayerState.Error;
   }, [currentState]);
 
   // eslint-disable-next-line no-warning-comments
   // TODO: let _io: IntersectionObserver | undefined;
-
-  function seek(value: number | string, nextState: PlayerState): void {
-    if (!dotLottiePlayer) return;
-    let frameValue = value;
-
-    if (typeof frameValue === 'number') {
-      frameValue = Math.round(frameValue);
-    }
-
-    // Extract frame number from either number or percentage value
-    const matches = /^(\d+)(%?)$/u.exec(frameValue.toString());
-
-    if (!matches) {
-      return;
-    }
-
-    // Calculate and set the frame number
-    const nextFrame = matches[2] === '%' ? (dotLottiePlayer.totalFrames * Number(matches[1])) / 100 : matches[1];
-
-    // Set seeker to new frame number
-    if (nextFrame === undefined) return;
-    // Send lottie player to the new frame
-    if (nextState === PlayerState.Playing) {
-      dotLottiePlayer.goToAndPlay(nextFrame, true);
-    } else {
-      dotLottiePlayer.goToAndStop(nextFrame, true);
-      dotLottiePlayer.pause();
-    }
-  }
-
-  function handleSeekChange(event: React.FormEvent<HTMLInputElement>): void {
-    if (!dotLottiePlayer || !Number(event.currentTarget.value)) {
-      return;
-    }
-    const newFrame: number = (Number(event.currentTarget.value) / 100) * dotLottiePlayer.totalFrames;
-
-    seek(newFrame, currentState);
-  }
-
-  function togglePlay(): void {
-    if (!dotLottiePlayer) return;
-    if (currentState === PlayerState.Playing) {
-      dotLottiePlayer.pause();
-    } else {
-      dotLottiePlayer.play();
-    }
-  }
-
-  function setLooping(value: boolean): void {
-    if (dotLottiePlayer) {
-      setIsLoop(value);
-      dotLottiePlayer.setLoop(value);
-    }
-  }
 
   useEffect(() => {
     dotLottiePlayer?.updateSrc(src);
@@ -264,106 +207,49 @@ export const DotLottiePlayer: React.FC<DotLottiePlayerProps> = ({
     }
   }, [currentState]);
 
-  function renderControls(): JSX.Element {
-    return (
-      <div aria-label="lottie-animation-controls" className="toolbar">
-        <button
-          onClick={togglePlay}
-          className={`${isPlaying || isPaused ? 'active' : ''}`}
-          style={{ alignItems: 'center' }}
-          // tabindex={0}
-          aria-label="play-pause"
-        >
-          {isPlaying ? (
-            <svg width="24" height="24" aria-hidden="true" focusable="false">
-              <path d="M14.016 5.016H18v13.969h-3.984V5.016zM6 18.984V5.015h3.984v13.969H6z" />
-            </svg>
-          ) : (
-            <svg width="24" height="24" aria-hidden="true" focusable="false">
-              <path d="M8.016 5.016L18.985 12 8.016 18.984V5.015z" />
-            </svg>
-          )}
-        </button>
-        <button
-          onClick={(): void => dotLottiePlayer?.stop()}
-          className={`${isStopped ? 'active' : ''}`}
-          style={{ alignItems: 'center' }}
-          aria-label="stop"
-        >
-          <svg width="24" height="24" aria-hidden="true" focusable="false">
-            <path d="M6 6h12v12H6V6z" />
-          </svg>
-        </button>
-        <input
-          className="seeker"
-          type="range"
-          min={0}
-          step={0}
-          max={100}
-          value={seeker || 0}
-          onInput={(event): void => handleSeekChange(event)}
-          onMouseDown={(): void => {
-            setPrevState(currentState);
-            dotLottiePlayer?.freeze();
-          }}
-          onMouseUp={(): void => {
-            seek(frame || 0, _prevState);
-          }}
-          aria-valuemin={1}
-          aria-valuemax={100}
-          role="slider"
-          aria-valuenow={seeker}
-          aria-label="lottie-seek-input"
-        />
-        <button
-          onClick={(): void => {
-            setLooping(!isLoop);
-          }}
-          className={isLoop ? 'active' : ''}
-          style={{ alignItems: 'center' }}
-          aria-label="loop-toggle"
-        >
-          <svg width="24" height="24" aria-hidden="true" focusable="false">
-            <path d="M17.016 17.016v-4.031h1.969v6h-12v3l-3.984-3.984 3.984-3.984v3h10.031zM6.984 6.984v4.031H5.015v-6h12v-3l3.984 3.984-3.984 3.984v-3H6.984z" />
-          </svg>
-        </button>
-      </div>
-    );
-  }
-
   return (
-    <div
-      className={`dotlottie-container main ${controls ? 'controls' : ''} ${className}`}
-      lang="en"
-      role="img"
-      {...(testId && {
-        'data-testid': testId,
-        'data-player-state': currentState,
-        'data-player-seeker': seeker || 0,
-      })}
-      {...props}
+    <DotLottieContext.Provider
+      value={{
+        currentState,
+        dotLottiePlayer,
+        frame,
+        seeker,
+        loop: Boolean(loop),
+      }}
     >
       <div
-        ref={container}
-        data-name="my-anim"
-        className={`animation ${controls ? 'controls' : ''}`}
-        style={{ background, position: 'relative' }}
+        className={`dotlottie-container main ${controls ? 'controls' : ''} ${className}`}
+        lang="en"
+        role="img"
         {...(testId && {
-          'data-testid': `animation`,
+          'data-testid': testId,
+          'data-player-state': currentState,
+          'data-player-seeker': seeker || 0,
         })}
+        {...props}
       >
-        {isError && (
-          <div
-            {...(testId && {
-              'data-testid': `error`,
-            })}
-            className="error"
-          >
-            ⚠️
-          </div>
-        )}
+        <div
+          ref={container}
+          data-name="my-anim"
+          className={`animation ${controls ? 'controls' : ''}`}
+          style={{ background, position: 'relative' }}
+          {...(testId && {
+            'data-testid': `animation`,
+          })}
+        >
+          {isError && (
+            <div
+              {...(testId && {
+                'data-testid': `error`,
+              })}
+              className="error"
+            >
+              ⚠️
+            </div>
+          )}
+        </div>
+        {children}
       </div>
-      {controls && renderControls()}
-    </div>
+    </DotLottieContext.Provider>
   );
 };
